backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Google Cloud
google-cloud-firestore==2.13.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.0
google-cloud-tasks==2.14.0
google-cloud-logging==3.8.0
google-cloud-kms==2.19.0
google-auth==2.23.0
firebase-admin==6.2.0

# Database
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
authlib==1.2.1
ldap3==2.9.1

# AI & ML
anthropic==0.7.0
openai==1.3.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0

# RPA & Automation
pyautogui==0.9.54
pynput==1.7.6
opencv-python==4.8.1
pytesseract==0.3.10
pillow==10.1.0
selenium==4.15.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
pdfplumber==0.10.3
openpyxl==3.1.2

# Payments
stripe==7.0.0

# Communications
twilio==8.10.0
sendgrid==6.10.0
pywebpush==1.14.0
websockets==12.0
aiosmtplib==3.0.1

# Utilities
jinja2==3.1.2
python-dateutil==2.8.2
croniter==2.0.1
celery==5.3.4
requests==2.31.0
aiohttp==3.9.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.0.0
responses==0.24.1
freezegun==1.2.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0
pylint==3.0.2
